MlxPlatformPkg: map normal memory as inner shareable
This fixes a whole bunch of shennanigans, including
completely unuseable/slow UEFI with any PCIe devices
enumerated:
- Glacial USB boot
- Glacial PXE boot.
- No working UEFI Shell.

The slowness was caused by disabling data cache as soon as
PCIe was initialized. This in turn was done because
PCIe is cache coherent, while UEFI was mapping normal
memory as NSH. A location in Normal memory with the Non-shareable
attribute does not require the hardware to make data accesses by
different observers coherent, unless the memory is Non-cacheable.

Blindly moving to ISH mappings causes a hang in PrePi. It turns
out one of the ISH mappings overlapped a secure memory
region. Hard to say why that would make the Lzma code never finish,
but perhaps there was some speculation involved...dunno.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>